CYA shortcuts + variable CYA answers
Added support for the variable ULN and Year of Completion answers in the CYA screen. They now pre-process which variable to show based on what the user has selected.

Added the CYA shortcuts in the authorisation journey. While iterating, I made two version of this with. I've stored the code for both in a new 'code snippets' folder, just in case that's useful to me/anyone in the future.